updated appy.py
Changes in 'def convertImage()' function (along with import base64)
NOTE : base64 encoding takes 8-bit binary byte data and encodes it uses only the characters A-Z, a-z, 0-9, +, /* so it can be transmitted over channels that do not preserve all 8-bits of data, such as email.Hence, it wants a string of 8-bit bytes. You create those in Python 3 with the b'' syntax.